% Auto-generated: do not edit by hand
\name{dashFacebookLogin}

\alias{dashFacebookLogin}

\title{DashFacebookLogin component}

\description{
DashFacebookLogin is a wrapper around the react-facebook-login component. For details on its configuration, see https://www.npmjs.com/package/react-facebook-login
}

\usage{
dashFacebookLogin(id=NULL, appId=NULL, size=NULL, scope=NULL, fields=NULL,
returnScopes=NULL, autoLoad=NULL, xfbml=NULL, cookie=NULL,
textButton=NULL, cssClass=NULL, redirectUri=NULL,
version=NULL, icon=NULL, language=NULL, isMobile=NULL,
isDisabled=NULL, tag=NULL, state=NULL, authType=NULL,
responseType=NULL, facebookLoginResponse=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{appId}{Character. The Facebook App ID.}

\item{size}{A value equal to: "small", "medium", "metro". }

\item{scope}{Character. }

\item{fields}{Character. }

\item{returnScopes}{Logical. }

\item{autoLoad}{Logical. }

\item{xfbml}{Logical. }

\item{cookie}{Logical. }

\item{textButton}{Character. }

\item{cssClass}{Character. }

\item{redirectUri}{Character. }

\item{version}{Character. }

\item{icon}{Character. }

\item{language}{Character. }

\item{isMobile}{Logical. }

\item{isDisabled}{Logical. }

\item{tag}{Character. }

\item{state}{Character. }

\item{authType}{Character. }

\item{responseType}{Character. }

\item{facebookLoginResponse}{Named list. The properties returned by the Facebook login callback}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

